Switched the algorithm to be part of the Logic component rather than the Model component
