+++ src/main/java/seedu/address/model/event/EventList.java
@@ -71,8 +71,8 @@ public class EventList {
     }
 
     /**
-     * Removes the equivalent event from the Schedule.
-     * The Event must exist in the Schedule.
+     * Removes the equivalent event from the Events Schedule.
+     * The Event must exist in the Events Schedule.
      * @param toRemove
      */
     public void remove(Event toRemove) {
